Remove redundant cast

+++ source/KissFFT.cpp
@@ -361,7 +361,7 @@ static void kf_factor(int n, int *facbuf)
 {
     int p = 4;
     double floor_sqrt;
-    floor_sqrt = floor(sqrt((double)n));
+    floor_sqrt = floor(sqrt(n));
 
     /*factor out powers of 4, powers of 2, then any remaining primes */
     do
